Cleaned up adapter
