+++ app/src/main/res/layout/task_list_item.xml
@@ -1,16 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent" android:layout_height="50dp"
-    android:id="@+id/taskItemLayout">
+    android:id="@+id/taskItemLayout"
+    android:layout_width="match_parent"
+    android:layout_height="50dp">
 
     <TextView
+        android:id="@+id/tvTaskName"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="15dp"
         android:layout_centerVertical="true"
-        android:id="@+id/task_name"
-        android:textSize="20dp"
+        android:layout_marginStart="15dp"
         android:textColor="@android:color/black"
-        />
+        android:textSize="20sp" />
 
 </RelativeLayout>
